% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_optimization_matrices.R
\name{build_optimization_matrices}
\alias{build_optimization_matrices}
\title{build_optimization_matrices}
\usage{
build_optimization_matrices(solution_space, iteration_count, lines,
  oltc.trigger, allowed_voltage = 0.03, verbose = 0)
}
\arguments{
\item{solution_space}{list of possible reinforcement measures for lines, transformator and parallel lines.}

\item{iteration_count}{counter for reinforcement iteration.}

\item{lines}{parameters of original grid assets.}

\item{oltc.trigger}{indication for OLTC transformator usage.}

\item{allowed_voltage}{limit of voltage deviation.}

\item{verbose}{Value greater than zero to display step by step of reinforcement.}
}
\value{
List of right hand, operatot and left hand side of optimization equations
}
\description{
This function creates the side conditions and objective function for the optimization problem.
}
\keyword{,}
\keyword{optimization}
\keyword{solution}
\keyword{space}
