% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_solution_space.R
\name{build_solution_space}
\alias{build_solution_space}
\title{build_solution_space}
\usage{
build_solution_space(grid, avail_asset_types, verbose = 0, line_types = NA,
  trafo_types = NA)
}
\arguments{
\item{grid}{List containing initial grid data.}

\item{avail_asset_types}{List of available line and trafo types for replacement}

\item{verbose}{Value greater than zero to display step by step of reinforcement}

\item{line_types}{database containing all line types and its specification}

\item{trafo_types}{database containing all line types and its specification}
}
\value{
Output is list containing 3 dataframe containing possible asset types in the grid and its specifications
}
\description{
function to collect possible asset types for lines, transformer and possible parallel lines
in the grid.
}
