% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_solution_space_A.R
\name{build_solution_space_A}
\alias{build_solution_space_A}
\title{build_solution_space_A}
\usage{
build_solution_space_A(grid, expansion_alternatives, verbose = 0)
}
\arguments{
\item{grid}{List containing initial grid data.}

\item{expansion_alternatives}{data frame containing available grid assets for either line or transformator reinforcement}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
Output is a dataframe containing possible cable types for each lines in the grid with its specifications
}
\description{
Function to create possible solution for grid reinforcement on lines
}
