% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trafo_I_b.R
\name{calc_trafo_I_b}
\alias{calc_trafo_I_b}
\title{calc_trafo_I_b}
\usage{
calc_trafo_I_b(grid, replacement_trafo, verbose = 3)
}
\arguments{
\item{grid}{List containing initial grid data.}

\item{replacement_trafo}{transformer to be installed in the grid}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
data frame containing transformer types and the current tha flows through it
}
\description{
funstion to calculate current flowing through the transformer in the grid
}
