% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_trafo_impedance.R
\name{calc_trafo_impedance}
\alias{calc_trafo_impedance}
\title{calc_trafo_impedance}
\usage{
calc_trafo_impedance(solution_space, Vref = 400)
}
\arguments{
\item{solution_space}{dataframe containing possible transformer types in the grid and its specifications}

\item{Vref}{reference voltage in the grid, default value is 400 V for low voltage distribution grid}
}
\value{
solution_space data with impedance
}
\description{
function that calculates the impedance of transformers
}
