% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_impedances.R
\name{calculate_impedances}
\alias{calculate_impedances}
\title{calculate_impedances}
\usage{
calculate_impedances(grid, solution_space, type)
}
\arguments{
\item{grid}{List containing initial grid data.}

\item{solution_space}{dataframe containing possible cable types in the grid and its specifications}

\item{type}{value is line or trafo to calculate impedance of respective asset type}
}
\value{
solution_space data with impedance
}
\description{
function to calculate impedances for the assets within the solution space.
}
