% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_voltage_drop_per_line.R
\name{calculate_voltage_drop_per_line}
\alias{calculate_voltage_drop_per_line}
\title{calculate_voltage_drop_per_line}
\usage{
calculate_voltage_drop_per_line(Vref, R, X, I)
}
\arguments{
\item{Vref}{voltage reference of the grid}

\item{R}{Resistance}

\item{X}{Impedance}

\item{I}{Current}
}
\value{
voltage drop of the specific lines or transformer
}
\description{
function to calculate the voltage drop per element (line or transformer) linear
}
