% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_number_parallel_lines.R
\name{check_number_parallel_lines}
\alias{check_number_parallel_lines}
\title{check_number_parallel_lines}
\usage{
check_number_parallel_lines(parallel_lines, lines)
}
\arguments{
\item{parallel_lines}{number of parallel lines created}

\item{lines}{data frame from grid data containing grid assets and its parameters}
}
\value{
warnings if the number of created parallel lines is not correct
}
\description{
Checks if the number of obtained parallel lines is correct
}
