% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reinforcement.R
\name{check_reinforcement}
\alias{check_reinforcement}
\title{check_reinforcement}
\usage{
check_reinforcement(lines, line_types = NA, trafo_types = NA)
}
\arguments{
\item{lines}{data frame from grid data containing grid assets and its parameters}

\item{line_types}{database containing all line types and its specification}

\item{trafo_types}{database containing all line types and its specification}
}
\description{
This function checks if reinforcement is possible
}
