% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_solution_space.R
\name{combine_solution_space}
\alias{combine_solution_space}
\title{combine_solution_space}
\usage{
combine_solution_space(solution_space, chosen_expansions, verbose = 0)
}
\arguments{
\item{solution_space}{dataframe containing possible cable types in the grid and its specifications}

\item{chosen_expansions}{vector of optimization result}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
This function returns a solution space in a single matrix containing only important parameters. 
Moreover the chose expansion alternatives are integrated
}
\description{
This function combines the the solution_spaces with chosen expansion from optimization result
}
