% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_one_measure_per_slot.R
\name{constraint_one_measure_per_slot}
\alias{constraint_one_measure_per_slot}
\title{constraint_one_measure_per_slot}
\usage{
constraint_one_measure_per_slot(solution_space, verbose = 0)
}
\arguments{
\item{solution_space}{dataframe containing possible cable types in the grid and its specifications}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
This function creates the side conditions and objective function for the optimization problem. 
The ouput is a list. That contains A, b and const.dir.
}
\description{
This function creates conditions to ensure that only one expansion measure 
per slot in the original grid is chosen
}
