% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.lines.R
\name{convert.lines}
\alias{convert.lines}
\title{check.lines}
\usage{
convert.lines(grid, verbose = 1)
}
\arguments{
\item{grid}{List containing information of grid to be optimized.}

\item{verbose}{Verbosity level. value greater than zero to display step by step of reinforcement}
}
\value{
grid data with transmission ratio and converted lines
}
\description{
prepares lines and trafos for load flow calculation
}
