% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_current_constrain_trafo.R
\name{create_current_constrain_trafo}
\alias{create_current_constrain_trafo}
\title{create_current_constrain_trafo}
\usage{
create_current_constrain_trafo(solution_space, big_M, verbose = 0)
}
\arguments{
\item{solution_space}{dataframe containing possible cable types in the grid and its specifications}

\item{big_M}{big M value for current limit optimization}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
This function creates the side conditions concernig the rated current for non parallel lines. 
The ouput is a list. That contains A, b and const.dir.
}
\description{
this function creates the current constrains for transformer in the grid
}
