% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_assets.R
\name{create_grid_assets}
\alias{create_grid_assets}
\title{create_grid_assets}
\usage{
create_grid_assets(grid, type = c("trafo", "line"))
}
\arguments{
\item{grid}{List, data frame containing information of grid to be optimized.}

\item{type}{select grid assets type to be generated, default is all asset types}
}
\value{
data frame with the grid assets combined with transmission ratio and current
}
\description{
the functions creates a data frame with the grid assets
}
