% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_one_expansion_constraint.R
\name{create_one_expansion_constraint}
\alias{create_one_expansion_constraint}
\title{create_one_expansion_constraint}
\usage{
create_one_expansion_constraint(solution_space, verbose = 0)
}
\arguments{
\item{solution_space}{dataframe containing possible cable types in the grid and its specifications}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
This function creates the side conditions and objective function for the optimization problem. 
The ouput is a list. That contains A, b and const.dir.
}
\description{
This function creates conditions to ensure that only one expansion measure 
is chosen for every part of the grid
}
