% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_voltage_constraints_parallel.R
\name{create_voltage_constraints_parallel}
\alias{create_voltage_constraints_parallel}
\title{create_voltage_constraints_parallel}
\usage{
create_voltage_constraints_parallel(solution_space, big_M, allowed_voltage,
  verbose = 0)
}
\arguments{
\item{solution_space}{dataframe containing possible cable and transformer types in the grid and its specifications}

\item{big_M}{big M value for voltage limit optimization}

\item{allowed_voltage}{allowed voltage deviatipn limit}

\item{verbose}{Value greater than zero to display step by step of reinforcement.}
}
\value{
This function creates the side conditions and objective function for the optimization problem. 
The ouput is a list. That contains A, b and c.
}
\description{
creates the constraints given by the voltage over the existing line, 
when a parallel line is added next to it.
}
