% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_leaves.R
\name{find_leaves}
\alias{find_leaves}
\title{find_leaves}
\usage{
find_leaves(lines, slack = "GRID")
}
\arguments{
\item{lines}{lines_data frame of package SimTOOL}

\item{slack}{Slack node of the grid.}
}
\value{
Names of leave nodes of the grid.
}
\description{
find the leaf node(s) and remove slack node of a grid. It searches the nodes having only one connection.
}
