% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_parallel_lines_update.R
\name{find_parallel_lines_update}
\alias{find_parallel_lines_update}
\title{find_parallel_lines}
\usage{
find_parallel_lines_update(lines, slack_node)
}
\arguments{
\item{lines}{lines_data frame of package SimTOOL}

\item{slack_node}{reference node that has has angular reference of 0 and voltage magnitude of 1 p.u.}
}
\value{
dataframe with possible begin and end nodes of parallel lines together with the line length.
}
\description{
Identifies possible locations for parallel lines for grid expansion. 
Starting point for all parallel lines is the busbar at the transformer. 
Possible end points are all nodes except leaf nodes.
}
