% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_paths_and_branches.R
\name{get_grid_paths_and_branches}
\alias{get_grid_paths_and_branches}
\title{get_grid_paths}
\usage{
get_grid_paths_and_branches(lines, slack_node = "GRID")
}
\arguments{
\item{lines}{lines_data frame of package SimTOOL}

\item{slack_node}{reference node that has has angular reference of 0 and voltage magnitude of 1 p.u.}
}
\value{
Output is a dataframe. Rows contain the lines. For each path a column indicates whether 
the corresponding line is contained (1) or not (0).
}
\description{
For each line the corresponding path in the grid is determined. It is based on
a shortest paths analysis from each end node of the grid to the low-voltage side of the transformer.
}
