% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_installation_cost.R
\name{get_installation_cost}
\alias{get_installation_cost}
\title{get_installation_cost}
\usage{
get_installation_cost(line, verbose = 0)
}
\arguments{
\item{line}{lines_data frame of package SimTOOL}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
installation cost of a specific segment. Material costs are in €/km in the costs data. 
Where installation costs are given in €/m.
}
\description{
calculate the installation cost for a line
}
