% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line_length.R
\name{get_line_length}
\alias{get_line_length}
\title{get_line_length}
\usage{
get_line_length(line_element, verbose = 0)
}
\arguments{
\item{line_element}{element value of line data frame}

\item{verbose}{Value greater than zero to display step by step of reinforcement}
}
\value{
line length in km
}
\description{
this function returns the line length from a lines$element
}
