% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.to.igraph.R
\name{grid.to.igraph}
\alias{grid.to.igraph}
\title{grid.to.igraph}
\usage{
grid.to.igraph(lines, directed = F)
}
\arguments{
\item{lines}{lines_data frame of package SimTOOL}

\item{directed}{logical giving if the graph shall be handled as. F is the default value and signifies undirected.}
}
\value{
igraph graph object
}
\description{
makes an igraph out of grid$lines. as result various functions of the 
large package "igraph" are applicable. e.g. different plots or shortest path calculations
}
\details{
possible extensions:
solutions for the different data formats provided by igraph --> additional argument "format"
printing: new vertice shape for Grid and Trafo nodes (manual page 279+)
use argument layout to generate reproduceable plots (coordinates are recently random)
}
