% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grid_rein.R
\name{plot_grid_rein}
\alias{plot_grid_rein}
\title{plot_grid_rein}
\usage{
plot_grid_rein(grid, U_set = 400, vertex_label = T, edge_label = T,
  allowed_voltage = 0.03)
}
\arguments{
\item{grid}{List containing grid data.}

\item{U_set}{lower voltage level in the grid}

\item{vertex_label}{logical to show node names and its voltage deviation}

\item{edge_label}{logical to show line types and its usage}

\item{allowed_voltage}{allowed voltage deviation limit}
}
\value{
plotted graph
}
\description{
create a visual graph of the grid along with each line's usage and each node's voltage deviation
}
