% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_power_control.R
\name{reactive_power_control}
\alias{reactive_power_control}
\title{reactive_power_control}
\usage{
reactive_power_control(lines, S_cal, verbose = 0)
}
\arguments{
\item{lines}{lines data of the grid}

\item{S_cal}{omplex vector giving all powers in internal notation for cal_nodes}

\item{verbose}{verbosity level}
}
\value{
This function change the value of active and reactive power of the feed-in. 
The ouput is new complex number of feed-in power.
}
\description{
change feed-in power based on specific power factor
}
