% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement.R
\name{reinforcement}
\alias{reinforcement}
\title{Reinforcement}
\usage{
reinforcement(grid, reinforcement_method, avail_asset_types = list(line =
  c("NAYY4x150", "2xNAYY4x240", "2xNAYY4x150", "4xNAYY4x240"), trafo =
  c("DOTEL_630")), verbose = 0)
}
\arguments{
\item{grid}{List containing initial grid data.}

\item{reinforcement_method}{Reinforcement method to be applied.}

\item{avail_asset_types}{List of available line and trafo types for replacement, default value line = "NAYY4x150", "2xNAYY4x240", "2xNAYY4x150", "4xNAYY4x240" 
trafo = "DOTEL_630", "DOTEL_1000".}

\item{verbose}{Verbosity level. value greater than zero to display step by step of reinforcement}
}
\value{
grid_solved
}
\description{
Simulate grid reinforcement and choose the optimum measures given the method, load and available assets data
}
