% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinforcement_iteration.R
\name{reinforcement_iteration}
\alias{reinforcement_iteration}
\title{Reinforcement Iteration}
\usage{
reinforcement_iteration(grid, method, avail_asset_types, U_set, iteration_count,
  oltc.trigger)
}
\arguments{
\item{grid}{List containing information of grid to be optimized.}

\item{method}{optimization method, either conventional or OLTC or RPC or both OLTC and RPC.}

\item{avail_asset_types}{List of available line and trafo types for replacement}

\item{U_set}{Grid's lower voltage level}

\item{iteration_count}{counter for reinforcement iteration.}

\item{oltc.trigger}{indication for OLTC transformator usage.}
}
\value{
grid_solved
}
\description{
build a reinforced grid based on available assets
}
