% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_line_types.R
\name{replace_line_types}
\alias{replace_line_types}
\title{replace_line_types}
\usage{
replace_line_types(lines, line_types = NA, verbose = 1)
}
\arguments{
\item{lines}{data frame from grid data containing grid assets and its parameters}

\item{line_types}{database containing all line types and its specification}

\item{verbose}{Verbosity level. value greater than zero to display step by step of reinforcement}
}
\value{
lines data frame with all parameter set
}
\description{
get line parameter set based on its type and database Data/types.RData
}
