% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_trafo_types.R
\name{replace_trafo_types}
\alias{replace_trafo_types}
\title{replace_trafo_types}
\usage{
replace_trafo_types(lines, trafo_types = NA, verbose = 0)
}
\arguments{
\item{lines}{data frame from grid data containing grid assets and its parameters}

\item{trafo_types}{database containing all transformer types and its specification}

\item{verbose}{Verbosity level. value greater than zero to display step by step of reinforcement}
}
\value{
lines data frame with all parameter set
}
\description{
get transformer parameter set based on its type and database Data/types.RData
}
