% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.prepare.grid.R
\name{wrapper.prepare.grid}
\alias{wrapper.prepare.grid}
\title{wrapper.prepare.grid}
\usage{
wrapper.prepare.grid(grid, check = F, U_set = NULL, oltc.trigger = F,
  verbose = 0)
}
\arguments{
\item{grid}{List containing information of grid to be optimized.}

\item{check}{logical, to check if the structure of the grid allows optimization}

\item{U_set}{Grid's lower voltage level}

\item{oltc.trigger}{indication for OLTC transformator usage.}

\item{verbose}{Verbosity level. value greater than zero to display step by step of reinforcement}
}
\value{
grid after load flow calculation
}
\description{
runs all necessary functions for preparing and execute loa flow calculation from SimTOOL.
}
